patch %p problem va_arg with unsigned long

+++ print_p/print_p.c
@@ -6,7 +6,7 @@
 /*   By: eriling <eriling@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/19 09:34:25 by eriling           #+#    #+#             */
-/*   Updated: 2021/01/20 15:13:44 by eriling          ###   ########.fr       */
+/*   Updated: 2021/01/21 09:31:35 by eriling          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ void 	print_p(t_param *param, va_list arg, t_param_len *len_printf)
 	int		space;
 	int		zero_precision;
 
-	res = ft_itoa_base((unsigned int)va_arg(arg, int), 16, 'a');
+	res = ft_itoa_base((unsigned long)va_arg(arg, long), 16, 'a');
 	len = ft_strlen(res);
 	var_init_int(param, &space, &zero_precision, len);
 	if (ft_strcmp("0", res) == 0)
